feat(refactor): Restructure views and utilities for improved caching and authentication handling
