+++ movie/movieapp/utils/__init__.py
@@ -1,0 +1,1 @@
+from .utils import sync_tmdb_movies, TMDBUtils
