+++ movie/movieapp/views/__init__.py
@@ -1,0 +1,4 @@
+from .user_views import UserViewSet, RegisterView
+from .movie_views import MovieViewSet
+from .favorite_movie_views import FavoriteMovieViewSet
+from .auth_views import token_obtain_pair, token_refresh
