+++ movie/movieapp/views/auth_views.py
@@ -1,0 +1,17 @@
+# movieapp/views/auth_views.py
+from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
+from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
+from movieapp.serializers import UserSerializer
+
+class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
+    def validate(self, attrs):
+        data = super().validate(attrs)
+        user = self.user
+        data['user'] = UserSerializer(user).data
+        return data
+
+class CustomTokenObtainPairView(TokenObtainPairView):
+    serializer_class = CustomTokenObtainPairSerializer
+
+token_obtain_pair = CustomTokenObtainPairView.as_view()
+token_refresh = TokenRefreshView.as_view()
