+++ movie/movieapp/views/mixins.py
@@ -1,4 +1,3 @@
-# movieapp/utils/cache_utils.py
 import logging
 from django.core.cache import cache
 from rest_framework.response import Response
